Moved 1856 Tracker functionality into G1856::Track
